Edit Entry is now connected and takes in ID of item that is selected
