+++ Screens/OverLimitEntries.js
@@ -2,10 +2,12 @@ import { View, Text, Button, StyleSheet } from "react-native";
 import React from "react";
 import EntriesList from "../Components/EntriesList";
 
-export default function OverLimitEntries({ navigation }) {
+export default function OverLimitEntries({ navigation, route }) {
+  // console.log(route.params);
+  // console.log("Over");
   return (
     <View style={styles.container}>
-      <EntriesList />
+      <EntriesList all={false} />
     </View>
   );
 }
